requirements.txt
requests==2.26.0
PyExecJS==1.5.1
pyinstaller==5.3